Add support for Brown Sauce schemas Refactor to allow support for multiple simultaneous schemas
