+++ regxmllib/src/main/java/com/sandflow/smpte/register/ElementsRegister.java
@@ -30,82 +30,42 @@ import com.sandflow.smpte.util.UL;
 import java.io.IOException;
 import java.io.Reader;
 import java.io.Writer;
-import java.util.ArrayList;
+import java.net.URI;
 import java.util.Collection;
 import java.util.HashMap;
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.Marshaller;
 import javax.xml.bind.Unmarshaller;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlElementWrapper;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlSeeAlso;
+import javax.xml.bind.annotation.XmlTransient;
 import javax.xml.bind.annotation.XmlType;
 
 /**
  * Elements Register as defined in SMPTE ST 335
  */
-@XmlRootElement(name = "ElementsRegister", namespace = ElementsRegister.XML_NAMESPACE)
-@XmlAccessorType(XmlAccessType.NONE)
-@XmlType(name = "")
-@XmlSeeAlso(value = ElementEntry.class)
-public class ElementsRegister {
-    
-    public final static String XML_NAMESPACE = "http://www.smpte-ra.org/schemas/335/2012";
+@XmlTransient
+public abstract class ElementsRegister {
 
-    private final HashMap<QualifiedSymbol, ElementEntry> entriesBySymbol = new HashMap<>();
+    private final HashMap<QualifiedSymbol, Entry> entriesBySymbol = new HashMap<>();
 
-    private final HashMap<UL, ElementEntry> entriesByUL = new HashMap<>();
+    private final HashMap<UL, Entry> entriesByUL = new HashMap<>();
 
-    @XmlElement(name = "Entry", namespace = XML_NAMESPACE)
-    @XmlElementWrapper(name = "Entries", namespace = XML_NAMESPACE)
-    private final ArrayList<ElementEntry> entries = new ArrayList<>();
-
-    public ElementsRegister() {
+    protected ElementsRegister() {
     }
 
-    public ElementEntry getEntryByUL(UL ul) {
+    public Entry getEntryByUL(UL ul) {
         return entriesByUL.get(ul);
     }
 
-    public void addEntry(ElementEntry entry) throws DuplicateEntryException {
-
-        QualifiedSymbol sym = new QualifiedSymbol(entry.getSymbol(), entry.getNamespaceName());
-
-        if (entriesByUL.containsKey(entry.getUL())) {
-            throw new DuplicateEntryException(
-                    String.format("UL = %s is already present (symbol = %s).",
-                            entry.getUL(),
-                            entry.getSymbol()
-                    )
-            );
-        }
-
-        if (entriesBySymbol.containsKey(sym)) {
-            throw new DuplicateEntryException(
-                    String.format(
-                            "Symbol = %s  is already present (UL = %s).",
-                            entry.getSymbol(),
-                            entry.getUL()
-                    )
-            );
-        }
-
-        entries.add(entry);
-        entriesByUL.put(entry.getUL(), entry);
-        entriesBySymbol.put(sym, entry);
+    public Entry getEntryBySymbol(QualifiedSymbol qs) {
+        return entriesBySymbol.get(qs);
     }
 
-    public Collection<ElementEntry> getEntries() {
-        return entries;
-    }
+    public abstract Collection<? extends Entry> getEntries();
 
     public void toXML(Writer writer) throws JAXBException, IOException {
 
-        JAXBContext ctx = JAXBContext.newInstance(ElementsRegister.class);
+        JAXBContext ctx = JAXBContext.newInstance(this.getClass());
 
         Marshaller m = ctx.createMarshaller();
         m.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT, true);
@@ -113,20 +73,81 @@ public class ElementsRegister {
         writer.close();
     }
 
-    public static ElementsRegister fromXML(Reader reader) throws JAXBException, IOException {
+    public static ElementsRegister fromXML(Reader reader) throws JAXBException, IOException, DuplicateEntryException {
 
-        JAXBContext ctx = JAXBContext.newInstance(ElementsRegister.class);
+        JAXBContext ctx = JAXBContext.newInstance(com.sandflow.smpte.register.catsup.ElementsRegisterModel.class);
 
         Unmarshaller m = ctx.createUnmarshaller();
         ElementsRegister reg = (ElementsRegister) m.unmarshal(reader);
 
-        for (ElementEntry te : reg.entries) {
+        for (Entry te : reg.getEntries()) {
             QualifiedSymbol sym = new QualifiedSymbol(te.getSymbol(), te.getNamespaceName());
+
+            if (reg.getEntryByUL(te.getUL()) != null) {
+                throw new DuplicateEntryException(
+                        String.format("UL = %s is already present (symbol = %s).",
+                                te.getUL(),
+                                te.getSymbol()
+                        )
+                );
+            }
+
+            if (reg.entriesBySymbol.get(sym) != null) {
+                throw new DuplicateEntryException(
+                        String.format(
+                                "Symbol = %s  is already present (UL = %s).",
+                                te.getSymbol(),
+                                te.getUL()
+                        )
+                );
+            }
+
             reg.entriesByUL.put(te.getUL(), te);
             reg.entriesBySymbol.put(sym, te);
         }
 
         return reg;
 
+    }
+
+    /**
+     * Single Entry in an Elements Register (SMPTE ST 335)
+     */
+    public static interface Entry {
+
+        String getApplications();
+
+        String getContextScope();
+
+        String getDefiningDocument();
+
+        String getDefinition();
+
+        Kind getKind();
+
+        String getName();
+
+        URI getNamespaceName();
+
+        String getNotes();
+
+        String getSymbol();
+
+        UL getTypeUL();
+
+        UL getUL();
+
+        String getUnitOfMeasure();
+
+        String getValueLength();
+
+        String getValueRange();
+
+        boolean isDeprecated();
+
+        @XmlType(name = "")
+        public static enum Kind {
+            NODE, LEAF
+        }
     }
 }
