+++ regxmllib/src/main/java/com/sandflow/smpte/register/GroupsRegister.java
@@ -30,71 +30,44 @@ import com.sandflow.smpte.util.UL;
 import java.io.IOException;
 import java.io.Reader;
 import java.io.Writer;
-import java.util.ArrayList;
+import java.net.URI;
 import java.util.Collection;
 import java.util.HashMap;
+import java.util.Set;
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.Marshaller;
 import javax.xml.bind.Unmarshaller;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlElementWrapper;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlSeeAlso;
+import javax.xml.bind.annotation.XmlTransient;
 import javax.xml.bind.annotation.XmlType;
 
 /**
  * Groups Register as defined in SMPTE ST 395
  */
-@XmlRootElement(name = "GroupsRegister", namespace = GroupsRegister.XML_NAMESPACE)
-@XmlAccessorType(XmlAccessType.NONE)
-@XmlType(name = "")
-@XmlSeeAlso(value = GroupEntry.class)
-public class GroupsRegister {
+@XmlTransient
+public abstract class GroupsRegister {
 
-    public final static String XML_NAMESPACE = "http://www.smpte-ra.org/schemas/395/2014";
 
-    private final HashMap<QualifiedSymbol, GroupEntry> entriesBySymbol = new HashMap<>();
+    private final HashMap<QualifiedSymbol, Entry> entriesBySymbol = new HashMap<>();
 
-    private final HashMap<UL, GroupEntry> entriesByUL = new HashMap<>();
-
-    @XmlElement(name = "Entry", namespace = XML_NAMESPACE)
-    @XmlElementWrapper(name = "Entries", namespace = XML_NAMESPACE)
-    private final ArrayList<GroupEntry> entries = new ArrayList<>();
+    private final HashMap<UL, Entry> entriesByUL = new HashMap<>();
 
     public GroupsRegister() {
     }
 
-    public GroupEntry getEntryByUL(UL ul) {
+    public Entry getEntryByUL(UL ul) {
         return entriesByUL.get(ul);
     }
 
-    public void addEntry(GroupEntry entry) throws DuplicateEntryException {
-
-        QualifiedSymbol sym = new QualifiedSymbol(entry.getSymbol(), entry.getNamespaceName());
-
-        if (entriesByUL.containsKey(entry.getUL())) {
-            throw new DuplicateEntryException(String.format("UL = %s is already present (symbol = %s).", entry.getUL(), entry.getSymbol()));
-        }
-
-        if (entriesBySymbol.containsKey(sym)) {
-            throw new DuplicateEntryException(String.format("Symbol = %s  is already present (UL = %s).", entry.getSymbol(), entry.getUL()));
-        }
-
-        entries.add(entry);
-        entriesByUL.put(entry.getUL(), entry);
-        entriesBySymbol.put(sym, entry);
+    public Entry getEntryBySymbol(QualifiedSymbol qs) {
+        return entriesBySymbol.get(qs);
     }
 
-    public Collection<GroupEntry> getEntries() {
-        return entries;
-    }
+    public abstract Collection<? extends Entry> getEntries();
 
     public void toXML(Writer writer) throws JAXBException, IOException {
 
-        JAXBContext ctx = JAXBContext.newInstance(GroupsRegister.class);
+        JAXBContext ctx = JAXBContext.newInstance(this.getClass());
 
         Marshaller m = ctx.createMarshaller();
         m.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT, true);
@@ -102,22 +75,107 @@ public class GroupsRegister {
         writer.close();
     }
 
-    public static GroupsRegister fromXML(Reader reader) throws JAXBException, IOException {
+    public static GroupsRegister fromXML(Reader reader) throws JAXBException, IOException, DuplicateEntryException {
 
-        JAXBContext ctx = JAXBContext.newInstance(GroupsRegister.class);
+        JAXBContext ctx = JAXBContext.newInstance(com.sandflow.smpte.register.brown_sauce.GroupsRegisterModel.class,
+                        com.sandflow.smpte.register.catsup.GroupsRegisterModel.class);
 
         Unmarshaller m = ctx.createUnmarshaller();
-
         GroupsRegister reg = (GroupsRegister) m.unmarshal(reader);
 
-        for (GroupEntry te : reg.entries) {
+        for (Entry te : reg.getEntries()) {
             QualifiedSymbol sym = new QualifiedSymbol(te.getSymbol(), te.getNamespaceName());
+
+            if (reg.getEntryByUL(te.getUL()) != null) {
+                throw new DuplicateEntryException(
+                        String.format("UL = %s is already present (symbol = %s).",
+                                te.getUL(),
+                                te.getSymbol()
+                        )
+                );
+            }
+
+            if (reg.entriesBySymbol.get(sym) != null) {
+                throw new DuplicateEntryException(
+                        String.format(
+                                "Symbol = %s  is already present (UL = %s).",
+                                te.getSymbol(),
+                                te.getUL()
+                        )
+                );
+            }
+
             reg.entriesByUL.put(te.getUL(), te);
             reg.entriesBySymbol.put(sym, te);
         }
 
         return reg;
 
+    }
+
+    /**
+     *
+     * @author pal
+     */
+    public static interface Entry {
+
+        String getApplications();
+
+        Collection<? extends Record> getContents();
+
+        String getDefiningDocument();
+
+        String getDefinition();
+
+        Kind getKind();
+
+        Set<Byte> getKlvSyntax();
+
+        String getName();
+
+        URI getNamespaceName();
+
+        String getNotes();
+
+        UL getParent();
+
+        String getSymbol();
+
+        UL getUL();
+
+        Boolean isConcrete();
+
+        boolean isDeprecated();
+
+
+        @XmlType(name = "")
+        public static enum Kind {
+            NODE, LEAF
+        }
+
+        /**
+         *
+         * @author pal
+         */
+        public static interface Record {
+
+            Boolean getDistinguished();
+
+            Boolean getIgnorable();
+
+            UL getItem();
+
+            Long getLimitLength();
+
+            Long getLocalTag();
+
+            Boolean getOptional();
+
+            Boolean getUniqueID();
+
+            String getValue();
+
+        }
     }
 
 }
